+++ controllers/user/dashboard.js
@@ -2,7 +2,7 @@
 const path = require('path');
 const express = require('express');
 
-const { requireUserAuth } = require('../../middleware/authed');
+const { requireUserAuth } = require('../../common/authed');
 const { injectCopy } = require('../../middleware/inject-content');
 
 const router = express.Router();
