+++ controllers/user/update-email.js
@@ -6,7 +6,7 @@ const express = require('express');
 const { Users } = require('../../db/models');
 const { redirectForLocale } = require('../../common/urls');
 const { csrfProtection } = require('../../common/cached');
-const { requireUserAuth } = require('../../middleware/authed');
+const { requireUserAuth } = require('../../common/authed');
 const { injectCopy } = require('../../middleware/inject-content');
 
 const logger = require('../../common/logger').child({ service: 'user' });
